code: split ui files and simplify dictionary window by using Gtk.Box instead of Grid
