--- zhudi/gui.py
@@ -1,743 +1,0 @@
-from gi import require_version
-
-require_version("Gtk", "4.0")
-require_version("Adw", "1")
-from gi.repository import Gtk, Adw, Pango, Gdk, GLib
-import re
-
-
-from zhudi import chinese_table
-from zhudi.processing import DictionaryTools, SegmentationTools
-
-
-CANGJIE5_OBJ = chinese_table.Cangjie5Table()
-ARRAY30_OBJ = chinese_table.Array30Table()
-WUBI86_OBJ = chinese_table.Wubi86Table()
-
-
-class DictionaryWidgetMain(object):
-    """Dictionary tab gui."""
-
-    def __init__(self, data_object):
-        self.data_object = data_object
-        self.language = ""
-        self.results_list = []
-        self.results_tree = None
-        self.search_field = None
-        self.chinese_label = None
-        self.translation_box = None
-
-    def build(self) -> Gtk.Grid:
-        # Search field
-        search_field = Gtk.Entry()
-        search_field.set_visible(True)
-        search_field.connect("activate", self.search_asked)
-        search_field.set_placeholder_text("Looking for something?")
-        self.search_field = search_field
-
-        # Go, search! button
-        go_button = Gtk.Button(label="Search")
-        go_button.connect("clicked", self.search_asked)
-
-        # Search + button box
-        sb_box = Gtk.Grid()
-        sb_box.attach(search_field, 0, 0, 5, 1)
-        sb_box.attach_next_to(go_button, search_field, Gtk.PositionType.RIGHT, 1, 1)
-        sb_box.set_column_homogeneous(True)
-        sb_box.set_column_spacing(10)
-
-        # Search label zone
-        frame_search = Gtk.Frame()
-        frame_search.set_child(sb_box)
-
-        # Results part in a list
-        self.results_list = Gtk.ListStore(str)
-        results_tree = Gtk.TreeView(model=self.results_list)
-        renderer = Gtk.CellRendererText()
-        results_tree.tvcolumn = Gtk.TreeViewColumn("Results", renderer, text=0)
-        results_tree.append_column(results_tree.tvcolumn)
-        self.results_list.cell = Gtk.CellRendererText()
-        results_tree.tvcolumn.pack_start(self.results_list.cell, True)
-        results_tree.set_enable_search(False)
-        results_tree.tvcolumn.set_sort_column_id(-1)
-        results_tree.set_reorderable(False)
-        self.results_tree = results_tree
-        select = results_tree.get_selection()
-        select.connect("changed", self.display_another_result)
-
-        results_scroll = Gtk.ScrolledWindow()
-        # No horizontal bar, automatic vertical bar
-        results_scroll.set_policy(Gtk.PolicyType.NEVER, Gtk.PolicyType.AUTOMATIC)
-        results_scroll.set_child(results_tree)
-
-        frame_results = Gtk.Frame()
-        frame_results.set_child(results_scroll)
-        frame_results.set_vexpand(True)
-
-        # Translation Label
-        translation_label = Gtk.Label()
-        translation_label.set_text("<big>Translation</big>")
-        translation_label.set_use_markup(True)
-
-        # Chinese label
-        self.chinese_label = Gtk.Label()
-        self.chinese_label.set_selectable(True)
-        self.chinese_label.set_focus_on_click(False)
-        # self.chinese_label.set_track_visited_links(False)
-        self.chinese_label.connect("activate-link", self.on_character_clicked)
-
-        # Translation view
-        self.translation_box = Gtk.TextView()
-        self.translation_box.set_editable(False)
-        self.translation_box.set_cursor_visible(False)
-        self.translation_box.set_wrap_mode(Gtk.WrapMode.WORD)
-
-        # No horizontal bar, vertical bar if needed
-        translation_scroll = Gtk.ScrolledWindow()
-        translation_scroll.set_child(self.translation_box)
-        translation_scroll.set_vexpand(True)
-
-        # Mapping of the main window
-        left_vertical_box = Gtk.Grid()
-        left_vertical_box.attach(frame_search, 0, 0, 1, 1)
-        left_vertical_box.attach_next_to(
-            frame_results, frame_search, Gtk.PositionType.BOTTOM, 1, 1
-        )
-        left_vertical_box.set_row_spacing(10)
-
-        right_vertical_box = Gtk.Grid()
-        right_vertical_box.attach(self.chinese_label, 0, 0, 1, 1)
-        right_vertical_box.attach_next_to(
-            translation_scroll, self.chinese_label, Gtk.PositionType.BOTTOM, 1, 1
-        )
-        right_vertical_box.set_column_homogeneous(True)
-
-        frame_translation = Gtk.Frame()
-        frame_translation.set_label_widget(translation_label)
-        frame_translation.set_child(right_vertical_box)
-
-        horizontal_box = Gtk.Grid()
-        horizontal_box.attach(left_vertical_box, 0, 0, 1, 1)
-        horizontal_box.attach_next_to(
-            frame_translation, left_vertical_box, Gtk.PositionType.RIGHT, 1, 1
-        )
-        horizontal_box.set_column_homogeneous(True)
-        horizontal_box.set_column_spacing(10)
-        return horizontal_box
-
-    def on_character_clicked(self, chinese_label, href) -> None:
-        self.search_field.set_text(href)
-        self.search_asked(self.search_field)
-
-    def search_asked(self, search_field: Gtk.Entry) -> None:
-        """Start search when users hit ENTER or the search button."""
-        search_field.grab_focus()
-        text = search_field.get_text()
-        DICTIONARY_TOOLS_OBJECT.reset_search()
-        if text == "":
-            self.results_list.clear()
-            self.display_translation(0)
-        else:
-            self.language = self.determine_language(text)
-            if self.language == "Latin":
-                DICTIONARY_TOOLS_OBJECT.search(self.data_object.translation, text)
-                DICTIONARY_TOOLS_OBJECT.search(self.data_object.pinyin, text)
-            elif self.data_object.hanzi == "Traditional":
-                DICTIONARY_TOOLS_OBJECT.search(self.data_object.traditional, text)
-            else:
-                DICTIONARY_TOOLS_OBJECT.search(self.data_object.simplified, text)
-            DICTIONARY_TOOLS_OBJECT.finish_search(self.data_object)
-            self.update_results()
-            self.results_tree.grab_focus()
-            self.display_translation(0)
-        self.results_tree.scroll_to_cell([0])
-
-    # end of search_asked
-
-    @staticmethod
-    def determine_language(input_text):
-        """
-        Determine the language of the input text, according to its content
-
-        """
-
-        if SEGMENTATION_TOOLS_OBJECT.is_not_chinese(input_text):
-            return "Latin"
-        else:
-            return "Chinese"
-
-    def display_translation(self, which):
-        """Handles the display of the translation for the selected element."""
-
-        if len(DICTIONARY_TOOLS_OBJECT.index) == 0:
-            self.translation_box.get_buffer().set_text("")
-            if len(self.results_list) == 0:
-                self.results_list.append(["Nothing found."])
-            return
-        else:
-            index = DICTIONARY_TOOLS_OBJECT.index[which]
-
-        characters = self.data_object.get_chinese(index)
-
-        translation = re.sub(
-            r"\[(.*?)\]",
-            lambda x: "["
-            + DictionaryTools.romanizePinyin(self.data_object, x.group(1))
-            + "]",
-            self.data_object.translation[index],
-        )
-        numbered_translations = "".join(
-            f"{i+1}. {t}\n" for i, t in enumerate(translation.split("/"))
-        )
-
-        pronunciation_string = DictionaryTools.romanize(self.data_object, index)
-
-        # Display different writing methods for the entry
-        cangjie5_displayed = "".join(
-            f"[{CANGJIE5_OBJ.proceed(hanzi, self.data_object.cangjie5)[1]}]"
-            for hanzi in characters
-        )
-        array30_displayed = "".join(
-            f"[{ARRAY30_OBJ.proceed(hanzi, self.data_object.array30)[1]}]"
-            for hanzi in characters
-        )
-        wubi86_code = "".join(
-            f"[{WUBI86_OBJ.proceed(hanzi, self.data_object.wubi86)[0]}]"
-            for hanzi in characters
-        )
-
-        # Display in the Translation box
-        # The very tiny space is there so that clicking on a character searches for the correct character.
-        # Not sure why but it seems to help
-        self.chinese_label.set_markup(
-            '<span size="1pt"> </span>'.join(
-                f'<a href="{ch}"><span size="60pt" underline="none">{ch}</span></a>'
-                for ch in characters
-            )
-        )
-        translation_buffer = self.translation_box.get_buffer()
-        translation_buffer.set_text("")
-        translation_buffer.insert_markup(
-            iter=translation_buffer.get_end_iter(),
-            markup="<b>Pronunciation</b>\n<span foreground='#268bd2'>"
-            + pronunciation_string
-            + "</span>\n\n"
-            + "<b>Meaning</b>\n"
-            + GLib.markup_escape_text(numbered_translations, -1)
-            + "<b>Input methods codes</b>\n"
-            + "Array30 (行列30): \n"
-            + array30_displayed
-            + "\n\n"
-            + "Cangjie5 (倉頡5): \n"
-            + cangjie5_displayed
-            + "\n\n"
-            + "Wubi86 (五筆86): \n"
-            + wubi86_code,
-            len=-1,
-        )
-
-    def update_results(self):
-        """Clear, and refill the result list."""
-        self.results_list.clear()
-        displayed_index = 1
-        threashold = 40  # threshold for line wrap
-        for k in DICTIONARY_TOOLS_OBJECT.index:
-            if self.language == "latin":
-                string = self.data_object.translation[k]
-            elif self.data_object.hanzi == "traditional":
-                string = self.data_object.traditional[k]
-            else:
-                string = self.data_object.simplified[k]
-            if len(string) > threashold:
-                string = str(displayed_index) + ". " + string[0:threashold] + "…"
-            else:
-                string = str(displayed_index) + ". " + string
-            string = string[:-1]  # no \n
-            self.results_list.append([string])
-            displayed_index += 1
-
-    def display_another_result(self, selection):
-        """Display the newly selected result."""
-        model, treeiter = selection.get_selected()
-        if treeiter is not None:
-            self.display_translation(model[treeiter].path[0])
-
-
-class SegmentationWidget(object):
-    """Class that defines the segmentation GUI layer."""
-
-    def __init__(self, data_object):
-        self.frame_label = None
-        self.horizontal_separator = None
-        self.go_button = None
-        self.title_box = None
-        self.text_field = None
-        self.scrolledwindow = None
-        self.results_list = []
-        self.frame_results = None
-        self.left_vertical_box = None
-        self.right_vertical_box = None
-        self.results_label = None
-        self.results_field = None
-        self.results_scroll = None
-        self.results_frame = None
-        self.horizontal_box = None
-        self.data_object = data_object
-
-    def build(self):
-        """Mandatory GUI build method."""
-        # Frame label
-        self.frame_label = Gtk.Label()
-        self.frame_label.set_text("<big>Chinese text to process:</big>")
-        self.frame_label.set_use_markup(True)
-
-        # Horzontal separator
-        self.horizontal_separator = Gtk.Separator()
-
-        # Go! button
-        self.go_button = Gtk.Button(label="Go!")
-        self.go_button.connect(
-            "clicked", lambda x: self.go_segment(self.text_field.get_buffer())
-        )
-        # Frame title + Go! button
-        self.title_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=3)
-        self.title_box.append(self.frame_label)
-        self.title_box.append(self.horizontal_separator)
-        self.title_box.append(self.go_button)
-
-        # Text field (to process)
-        self.text_field = Gtk.TextView()
-        self.text_field.set_editable(True)
-        self.text_field.set_cursor_visible(True)
-        self.text_field.set_wrap_mode(Gtk.WrapMode.CHAR)
-        self.scrolledwindow = Gtk.ScrolledWindow()
-        self.scrolledwindow.set_hexpand(True)
-        self.scrolledwindow.set_vexpand(True)
-        self.scrolledwindow.set_child(self.text_field)
-
-        # Results part in a list
-        self.results_list = Gtk.ListStore(str)
-        results_tree = Gtk.TreeView(model=self.results_list)
-        renderer = Gtk.CellRendererText()
-        results_tree.tvcolumn = Gtk.TreeViewColumn("Results", renderer, text=0)
-        results_tree.append_column(results_tree.tvcolumn)
-        self.results_list.cell = Gtk.CellRendererText()
-        results_tree.tvcolumn.pack_start(self.results_list.cell, True)
-        results_tree.set_enable_search(False)
-        results_tree.tvcolumn.set_sort_column_id(-1)
-        results_tree.set_reorderable(False)
-        select = results_tree.get_selection()
-        select.connect("changed", self.word_selected)
-
-        results_scroll = Gtk.ScrolledWindow()
-        # No horizontal bar, automatic vertical bar
-        results_scroll.set_policy(Gtk.PolicyType.NEVER, Gtk.PolicyType.AUTOMATIC)
-        results_scroll.set_child(results_tree)
-
-        self.frame_results = Gtk.Frame()
-        self.frame_results.set_child(results_scroll)
-
-        # Mapping of window
-        self.left_vertical_box = Gtk.Grid()
-        # self.left_vertical_box.append(self.title_box)
-        self.left_vertical_box.attach_next_to(
-            self.scrolledwindow, self.title_box, Gtk.PositionType.BOTTOM, 1, 2
-        )
-        self.left_vertical_box.attach_next_to(
-            self.frame_results, self.scrolledwindow, Gtk.PositionType.BOTTOM, 1, 8
-        )
-        self.left_vertical_box.set_column_homogeneous(True)
-        self.left_vertical_box.set_row_homogeneous(True)
-
-        # Results frame
-        self.results_label = Gtk.Label()
-        self.results_label.set_text("<big>Translation</big>")
-        self.results_label.set_use_markup(True)
-        self.results_field = Gtk.TextView()
-        self.results_field.set_editable(False)
-        self.results_field.set_cursor_visible(False)
-        # No horizontal bar, vertical bar if needed
-        self.results_field.set_wrap_mode(Gtk.WrapMode.WORD)
-
-        self.results_scroll = Gtk.ScrolledWindow()
-        self.results_scroll.set_child(self.results_field)
-
-        self.results_frame = Gtk.Frame()
-        self.results_frame.set_label_widget(self.results_label)
-        self.results_frame.set_child(self.results_scroll)
-
-        self.right_vertical_box = Gtk.Box(
-            orientation=Gtk.Orientation.VERTICAL, spacing=2
-        )
-        self.right_vertical_box.append(self.results_frame)
-
-        self.horizontal_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=2)
-        self.horizontal_box.append(self.left_vertical_box)
-        self.horizontal_box.append(self.right_vertical_box)
-        self.horizontal_box.set_homogeneous(True)
-        return self.horizontal_box
-
-    def go_segment(self, text_buffer):
-        """Get the input text to segment, and display the words."""
-        beginning = text_buffer.get_start_iter()
-        end = text_buffer.get_end_iter()
-        # grab hidden characters
-        text = text_buffer.get_text(beginning, end, True)
-        text = text.replace(" ", "")
-        segmented_text = SEGMENTATION_TOOLS_OBJECT.sentence_segmentation(text)
-        self.display_results(segmented_text, text_buffer)
-        self.display_selectable_words(segmented_text)
-
-    def display_selectable_words(self, segmented_text):
-        """Add in the results the segmented words."""
-        widget = self.results_list
-        widget.clear()
-        for word in segmented_text:
-            widget.append([word])
-
-    @staticmethod
-    def display_results(text, results_buffer):
-        """Display the segmentation result directly in the input area.
-        This has a nice side effect: allowing you to copy the result.
-
-        """
-        text_to_display = ""
-        for item in text:
-            text_to_display += item
-            text_to_display += "    "
-        results_buffer.set_text(text_to_display)
-
-    def display_translation(self, index, bypass=False):
-        """Display the given index [of a word] in a nicely formatted output.
-        If bypass is True, then the index variable is a string that has to
-        be displayed as it.
-
-        """
-        translation_buffer = self.results_field.get_buffer()
-
-        if bypass:
-            translation_buffer.set_text(index)
-            return
-
-        if self.data_object.hanzi == "traditional":
-            hanzi_dic = self.data_object.traditional
-        else:
-            hanzi_dic = self.data_object.simplified
-
-        if self.data_object.romanisation == "zhuyin":
-            romanisation_dic = self.data_object.zhuyin
-        else:
-            romanisation_dic = self.data_object.pinyin
-
-        slash_list = []
-        trans_index = self.data_object.translation[index]
-        for line in range(len(trans_index)):
-            if trans_index[line] == "/":
-                slash_list.append(line)
-        temp = 0
-        trans = []
-        for key in range(len(slash_list)):
-            trans.append(str(key + 1) + ". " + trans_index[temp : slash_list[key]])
-            temp = slash_list[key] + 1
-        trans.append(
-            str(len(slash_list) + 1) + ". " + trans_index[temp : len(trans_index)]
-        )
-        string = ""
-        for local_index in range(len(slash_list) + 1):
-            string = string + trans[local_index] + "\n"
-
-        # Add [] around the pronunciation parts
-        p_string = romanisation_dic[index].split("/", 1)[0].split()
-        pronunciation_string = []
-        for point in range(len(p_string)):
-            if self.data_object.romanisation == "pinyin":
-                pronunciation_string.append(
-                    DictionaryTools.unicode_pinyin(p_string[point])
-                )
-                pronunciation_string.append(" ")
-            else:
-                pronunciation_string.append("[")
-                pronunciation_string.append(p_string[point])
-                pronunciation_string.append("]")
-        # Display the cangjie of the entry
-        cangjie5_displayed = ""
-        for hanzi in hanzi_dic[index]:
-            if hanzi != "\n":
-                key_code, displayed_code = CANGJIE5_OBJ.proceed(
-                    hanzi, self.data_object.cangjie5
-                )
-                cangjie5_displayed += "["
-                cangjie5_displayed += displayed_code
-                cangjie5_displayed += "]"
-        # Display the array30 of the entry
-        array30_displayed = ""
-        for hanzi in hanzi_dic[index]:
-            if hanzi != "\n":
-                key_code, displayed_code = ARRAY30_OBJ.proceed(
-                    hanzi, self.data_object.array30
-                )
-                array30_displayed += "["
-                array30_displayed += displayed_code
-                array30_displayed += "]"
-        # Display the array30 of the entry (here code = displayed)
-        wubi86_code = ""
-        for hanzi in hanzi_dic[index]:
-            if hanzi != "\n":
-                key_code, displayed_code = WUBI86_OBJ.proceed(
-                    hanzi, self.data_object.wubi86
-                )
-                wubi86_code += "["
-                wubi86_code += key_code
-                wubi86_code += "]"
-        # Display in the Translation box
-        translation_buffer.set_text(
-            "Chinese\n"
-            + hanzi_dic[index]
-            + "\n\n"
-            + "Pronunciation\n"
-            + "".join(pronunciation_string)
-            + "\n\n"
-            "Meaning\n"
-            + string
-            + "Input methods codes:\n"
-            + "Array30 (行列30): \n"
-            + array30_displayed
-            + "\n\n"
-            + "Cangjie5 (倉頡5): \n"
-            + cangjie5_displayed
-            + "\n\n"
-            + "Wubi86 (五筆86): \n"
-            + wubi86_code
-        )
-        bold = translation_buffer.create_tag(weight=Pango.Weight.BOLD)
-        big = translation_buffer.create_tag(size=30 * Pango.SCALE)
-        blue = translation_buffer.create_tag(foreground="#268bd2")
-
-        # "Chinese" in bold
-        start_1 = translation_buffer.get_iter_at_line(0)
-        end_1 = translation_buffer.get_iter_at_line(0)
-        end_1.forward_to_line_end()
-        translation_buffer.apply_tag(bold, start_1, end_1)
-
-        # Bigger Chinese
-        start_c = translation_buffer.get_iter_at_line(1)
-        end_c = translation_buffer.get_iter_at_line(1)
-        end_c.forward_to_line_end()
-        translation_buffer.apply_tag(big, start_c, end_c)
-
-        # "Pronunciation" in bold
-        start_2 = translation_buffer.get_iter_at_line(4)
-        end_2 = translation_buffer.get_iter_at_line(4)
-        end_2.forward_to_line_end()
-        translation_buffer.apply_tag(bold, start_2, end_2)
-
-        # "Pronunciation" in blue
-        start_3 = translation_buffer.get_iter_at_line(5)
-        end_3 = translation_buffer.get_iter_at_line(5)
-        end_3.forward_to_line_end()
-        translation_buffer.apply_tag(blue, start_3, end_3)
-
-        # "Meaning" in bold
-        start_3 = translation_buffer.get_iter_at_line(7)
-        end_3 = translation_buffer.get_iter_at_line(7)
-        end_3.forward_to_line_end()
-        translation_buffer.apply_tag(bold, start_3, end_3)
-        guess = string.count("\n")
-
-        # "Input methods codes" in bold
-        start_4 = translation_buffer.get_iter_at_line(guess + 7)
-        end_4 = translation_buffer.get_iter_at_line(guess + 7)
-        end_4.forward_to_line_end()
-        translation_buffer.apply_tag(bold, start_4, end_4)
-
-    def word_selected(self, selection):
-        """Display the selected word in the translation area."""
-        model, treeiter = selection.get_selected()
-        if treeiter is not None:
-            word = model[treeiter][0]
-            if word is not None:
-                index = SEGMENTATION_TOOLS_OBJECT.search_unique(word, self.data_object)
-                if index is None:
-                    self.display_translation(word, True)
-                else:
-                    self.display_translation(index)
-
-
-class OptionsWidget(object):
-    """Class defining the Options/About tab layout"""
-
-    def __init__(self, data_object):
-        self.data_object = data_object
-
-    def build(self) -> Gtk.Box:
-        vertical_box = Gtk.Box()
-        vertical_box.set_orientation(Gtk.Orientation.VERTICAL)
-        vertical_box.set_homogeneous(False)
-        vertical_box.set_spacing(10)
-
-        title = Gtk.Label(label="")
-        vertical_box.append(title)
-
-        horizontal_box = Gtk.Box()
-        label = Gtk.Label(label="Romanisation:")
-        horizontal_box.append(label)
-        romanization_dropdown = Gtk.DropDown.new_from_strings(["Pinyin", "Zhuyin"])
-        # Set active the saved value
-        if self.data_object.romanisation == "zhuyin":
-            romanization_dropdown.set_selected(1)
-        else:
-            romanization_dropdown.set_selected(0)
-        romanization_dropdown.connect("notify::selected", self.on_romanization_selected)
-        horizontal_box.append(romanization_dropdown)
-        # Empty space
-        space = Gtk.Label(label="")
-        horizontal_box.append(space)
-        horizontal_box.set_homogeneous(True)
-        vertical_box.append(horizontal_box)
-
-        horizontal_box = Gtk.Box()
-        label = Gtk.Label(label="Character set:")
-        horizontal_box.append(label)
-        characters_dropdown = Gtk.DropDown.new_from_strings(
-            ["Simplified", "Traditional"]
-        )
-        if self.data_object.hanzi == "traditional":
-            characters_dropdown.set_selected(1)
-        else:
-            characters_dropdown.set_selected(0)
-        characters_dropdown.connect("notify::selected", self.on_characters_selected)
-        horizontal_box.append(characters_dropdown)
-        horizontal_box.set_homogeneous(True)
-        space = Gtk.Label(label="")
-        horizontal_box.append(space)
-        vertical_box.append(horizontal_box)
-
-        about_text = Gtk.Frame()
-        about_text.set_label("Zhudi, 2011-2023")
-        about_text.set_label_align(1.0)
-        about_text.set_valign(Gtk.Align.END)
-        vertical_box.append(about_text)
-        about_text.set_vexpand(True)
-        return vertical_box
-
-    def on_romanization_selected(self, dropdown: Gtk.DropDown, _ignore) -> None:
-        selected = dropdown.get_selected_item().get_string()
-        if selected is not None:
-            self.data_object.romanisation = selected.lower()
-            self.data_object.save_config()
-
-    def on_characters_selected(self, dropdown: Gtk.DropDown, _ignore) -> None:
-        selected = dropdown.get_selected_item().get_string()
-        if selected is not None:
-            self.data_object.hanzi = selected.lower()
-            self.data_object.save_config()
-
-
-class MainWindow(Adw.ApplicationWindow):
-    """Class that defines the welcome screen, and gives access to other layers."""
-
-    def __init__(self, data_object, language, *args, **kwargs):
-        super(MainWindow, self).__init__(*args, **kwargs)
-        self.set_title("Zhudi")
-        self.set_default_size(700, 1000)
-        evk = Gtk.EventControllerKey.new()
-        evk.connect("key-pressed", self.on_key_press)
-        self.add_controller(evk)
-        self.data_object = data_object
-        self.language = language
-        self.dict_gui = None
-        self.dict_settings = None
-        self.seg_gui = None
-        self.tab_box = None
-        self.vbox = None
-
-    def build(self):
-        """Mandatory build function."""
-        self.data_object.create_set_chinese_characters()
-        global DICTIONARY_TOOLS_OBJECT
-        DICTIONARY_TOOLS_OBJECT = DictionaryTools()
-        global SEGMENTATION_TOOLS_OBJECT
-        SEGMENTATION_TOOLS_OBJECT = SegmentationTools()
-        SEGMENTATION_TOOLS_OBJECT.load(self.data_object)
-
-        # Tabs
-        self.dict_settings, self.dict_gui = self.dictionary_gui()
-        self.seg_gui = self.segmentation_gui()
-        self.options_gui = self.options_gui()
-
-        # Build the tab frame
-        self.tab_box = Gtk.Notebook()
-        self.tab_box.set_tab_pos(Gtk.PositionType.TOP)
-        self.tab_box.append_page(self.dict_gui, None)
-        self.tab_box.set_tab_label_text(self.dict_gui, "Dictionary")
-        self.tab_box.append_page(self.seg_gui, None)
-        self.tab_box.set_tab_label_text(self.seg_gui, "Segmentation")
-        self.tab_box.append_page(self.options_gui, None)
-        self.tab_box.set_tab_label_text(self.options_gui, "Options")
-
-        header = Gtk.HeaderBar()
-        box = Gtk.Box()
-        box.props.orientation = Gtk.Orientation.VERTICAL
-
-        box.append(header)
-        box.append(self.tab_box)
-
-        self.set_content(box)
-
-    def options_gui(self):
-        """Options tab."""
-        return OptionsWidget(self.data_object).build()
-
-    def dictionary_gui(self):
-        """Start the dictionary widget."""
-        ui = DictionaryWidgetMain(self.data_object)
-        ui.language = self.language
-        return ui, ui.build()
-
-    def segmentation_gui(self):
-        """Start the segmentation widget."""
-        return SegmentationWidget(self.data_object).build()
-
-    def on_key_press(self, event, keyval, keycode, state):
-        # Stop with Ctrl-w
-        if keyval == Gdk.KEY_w and state & Gdk.ModifierType.CONTROL_MASK:
-            self.close()
-
-        # Do nothing if random key
-        if isinstance(event, Gtk.EventControllerKey):
-            return
-
-        # Navigate in the result list with arrows and select the items
-        if self.tab_box.get_current_page() == 0:
-            search_key = (
-                len(event.string) > 0 and ord(event.string[0]) >= 0x20
-            ) or event.keyval in {Gdk.KEY_Left, Gdk.KEY_Right, Gdk.KEY_BackSpace}
-            if not self.dict_settings.search_field.has_focus() and search_key:
-                if event.keyval == Gdk.KEY_Left or event.keyval == Gdk.KEY_Right:
-                    self.dict_settings.search_field.grab_focus_without_selecting()
-                else:
-                    self.dict_settings.search_field.grab_focus()
-            elif event.keyval in {
-                Gdk.KEY_Up,
-                Gdk.KEY_Down,
-                Gdk.KEY_Page_Up,
-                Gdk.KEY_Page_Down,
-            }:
-                self.dict_settings.results_tree.grab_focus()
-
-
-class ZhudiApplication(Adw.Application):
-    def __init__(self, data_object, language):
-        super().__init__(application_id="com.github.jiehong.zhudi")
-        # self.connect('activate', self.on_activate)
-        self.data_object = data_object
-        self.language = language
-
-    def do_activate(self):
-        win = self.props.active_window
-        if not win:
-            win = MainWindow(
-                data_object=self.data_object, language=self.language, application=self
-            )
-            win.build()
-        win.present()
